pyproject.toml
[project]
name = "skycalc"
version = "0.1.0"
description = "Interactive airmass plotter for astronomical observations"
requires-python = ">=3.10"
dependencies = [
    "astropy>=6.0",
    "astroplan>=0.10",
    "dash>=2.14",
    "dash-bootstrap-components>=1.5",
    "plotly>=5.18",
    "pandas>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
skycalc = "skycalc.run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["skycalc"]
